pdf_to_jpeg/pyproject.toml
[project]
name = "pdf-to-jpeg"
version = "0.1.0"
description = "A CLI tool to convert PDF files to JPEG images"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "typer>=0.16.0",
]
